courses/api/serializers.py
from rest_framework import serializers

from courses.models import Subject, Module


class SubjectSerializer(serializers.ModelSerializer):
    class Meta:
        model = Subject
        fields = ['id', 'title', 'slug']


class ModuleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Module
        fields = ['id', 'title', 'description', 'order']


class CourseSerializer(serializers.ModelSerializer):
    modules = ModuleSerializer(many=True, read_only=True)

    class Meta:
        model = Module
        fields = ['id', 'title', 'overview', 'subject', 'slug', 'created', 'owner', 'modules']
