requirements.txt
Django
psycopg2
Pillow
django-widget-tweaks
django-braces
django-embed-video
python-memcached
django-memcache-status
djangorestframework
django-debug-toolbar